requirements.txt
python-docx
pillow
pydantic
openai